generic: add checks for zoned block device
Modify generic tests to require non-zoned block device

generic/108 is disabled on zoned block device because the LVM device
not always aligned to the zone boundary.

generic/471 is disabled because we cannot enable NoCoW on zoned
btrfs.

generic/570 is disabled because swap file which require nocow is not
usable on zoned btrfs.

Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Eryu Guan <guaneryu@gmail.com>